Move clique network into network
