+++ finality-aleph/src/lib.rs
@@ -44,10 +44,8 @@ mod nodes;
 mod party;
 mod session;
 mod session_map;
-mod tcp_network;
 #[cfg(test)]
 pub mod testing;
-mod validator_network;
 
 pub use abft::{Keychain, NodeCount, NodeIndex, Recipient, SignatureSet, SpawnHandle};
 pub use aleph_primitives::{AuthorityId, AuthorityPair, AuthoritySignature};
