+++ finality-aleph/src/network/clique/mock.rs
@@ -1,32 +1,168 @@
 use std::{
-    collections::{BTreeMap, HashMap, HashSet},
+    collections::HashMap,
+    fmt::{Display, Error as FmtError, Formatter},
     io::Result as IoResult,
     pin::Pin,
     task::{Context, Poll},
 };
 
-use codec::{Decode, Encode, Output};
+use codec::{Decode, Encode};
 use futures::{
     channel::{mpsc, oneshot},
     StreamExt,
 };
 use log::info;
-use rand::{thread_rng, Rng};
-use sc_service::{SpawnTaskHandle, TaskManager};
-use tokio::{
-    io::{duplex, AsyncRead, AsyncWrite, DuplexStream, ReadBuf},
-    runtime::Handle,
-    time::{error::Elapsed, interval, timeout, Duration},
+use rand::Rng;
+use tokio::io::{duplex, AsyncRead, AsyncWrite, DuplexStream, ReadBuf};
+
+use crate::network::{
+    clique::{
+        ConnectionInfo, Dialer, Listener, Network, PeerAddressInfo, PublicKey, SecretKey,
+        Splittable, LOG_TARGET,
+    },
+    mock::Channel,
+    AddressingInformation, Data, NetworkIdentity, PeerId,
 };
 
-use crate::{
-    network::{mock::Channel, AddressingInformation, Data, NetworkIdentity},
-    validator_network::{
-        mock::{key, random_keys, MockPublicKey, MockSecretKey},
-        ConnectionInfo, Dialer as DialerT, Listener as ListenerT, Network, PeerAddressInfo,
-        SecretKey, Service, Splittable,
-    },
-};
+/// A mock secret key that is able to sign messages.
+#[derive(Debug, PartialEq, Eq, Clone, Hash)]
+pub struct MockSecretKey([u8; 4]);
+
+/// A mock public key for verifying signatures.
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Hash, Encode, Decode)]
+pub struct MockPublicKey([u8; 4]);
+
+impl Display for MockPublicKey {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), FmtError> {
+        write!(f, "PublicKey({:?})", self.0)
+    }
+}
+
+impl AsRef<[u8]> for MockPublicKey {
+    fn as_ref(&self) -> &[u8] {
+        self.0.as_ref()
+    }
+}
+
+/// A mock signature, able to discern whether the correct key has been used to sign a specific
+/// message.
+#[derive(Debug, PartialEq, Eq, Clone, Hash, Encode, Decode)]
+pub struct MockSignature {
+    message: Vec<u8>,
+    key_id: [u8; 4],
+}
+
+impl PublicKey for MockPublicKey {
+    type Signature = MockSignature;
+
+    fn verify(&self, message: &[u8], signature: &Self::Signature) -> bool {
+        (message == signature.message.as_slice()) && (self.0 == signature.key_id)
+    }
+}
+
+impl PeerId for MockPublicKey {}
+
+#[async_trait::async_trait]
+impl SecretKey for MockSecretKey {
+    type Signature = MockSignature;
+    type PublicKey = MockPublicKey;
+
+    async fn sign(&self, message: &[u8]) -> Self::Signature {
+        MockSignature {
+            message: message.to_vec(),
+            key_id: self.0,
+        }
+    }
+
+    fn public_key(&self) -> Self::PublicKey {
+        MockPublicKey(self.0)
+    }
+}
+
+/// Create a random key pair.
+pub fn key() -> (MockPublicKey, MockSecretKey) {
+    let secret_key = MockSecretKey(rand::random());
+    (secret_key.public_key(), secret_key)
+}
+
+/// Create a HashMap with public keys as keys and secret keys as values.
+pub fn random_keys(n_peers: usize) -> HashMap<MockPublicKey, MockSecretKey> {
+    let mut result = HashMap::with_capacity(n_peers);
+    while result.len() < n_peers {
+        let (pk, sk) = key();
+        result.insert(pk, sk);
+    }
+    result
+}
+
+/// A mock that can be split into two streams.
+pub struct MockSplittable {
+    incoming_data: DuplexStream,
+    outgoing_data: DuplexStream,
+}
+
+impl MockSplittable {
+    /// Create a pair of mock splittables connected to each other.
+    pub fn new(max_buf_size: usize) -> (Self, Self) {
+        let (in_a, out_b) = duplex(max_buf_size);
+        let (in_b, out_a) = duplex(max_buf_size);
+        (
+            MockSplittable {
+                incoming_data: in_a,
+                outgoing_data: out_a,
+            },
+            MockSplittable {
+                incoming_data: in_b,
+                outgoing_data: out_b,
+            },
+        )
+    }
+}
+
+impl AsyncRead for MockSplittable {
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &mut ReadBuf<'_>,
+    ) -> Poll<IoResult<()>> {
+        Pin::new(&mut self.get_mut().incoming_data).poll_read(cx, buf)
+    }
+}
+
+impl AsyncWrite for MockSplittable {
+    fn poll_write(self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &[u8]) -> Poll<IoResult<usize>> {
+        Pin::new(&mut self.get_mut().outgoing_data).poll_write(cx, buf)
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<IoResult<()>> {
+        Pin::new(&mut self.get_mut().outgoing_data).poll_flush(cx)
+    }
+
+    fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<IoResult<()>> {
+        Pin::new(&mut self.get_mut().outgoing_data).poll_shutdown(cx)
+    }
+}
+
+impl ConnectionInfo for MockSplittable {
+    fn peer_address_info(&self) -> PeerAddressInfo {
+        String::from("MOCK_ADDRESS")
+    }
+}
+
+impl ConnectionInfo for DuplexStream {
+    fn peer_address_info(&self) -> PeerAddressInfo {
+        String::from("MOCK_ADDRESS")
+    }
+}
+
+impl Splittable for MockSplittable {
+    type Sender = DuplexStream;
+    type Receiver = DuplexStream;
+
+    fn split(self) -> (Self::Sender, Self::Receiver) {
+        (self.outgoing_data, self.incoming_data)
+    }
+}
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash, Encode, Decode)]
 pub struct MockAddressingInformation {
@@ -292,11 +428,9 @@ impl Splittable for UnreliableSplittable {
 }
 
 type Address = u32;
-type Addresses = HashMap<MockPublicKey, Address>;
+pub type Addresses = HashMap<MockPublicKey, Address>;
 type Callers = HashMap<MockPublicKey, (MockDialer, MockListener)>;
 type Connection = UnreliableSplittable;
-
-const TWICE_MAX_DATA_SIZE: usize = 32 * 1024 * 1024;
 
 #[derive(Clone)]
 pub struct MockDialer {
@@ -306,7 +440,7 @@ pub struct MockDialer {
 }
 
 #[async_trait::async_trait]
-impl DialerT<Address> for MockDialer {
+impl Dialer<Address> for MockDialer {
     type Connection = Connection;
     type Error = std::io::Error;
 
@@ -324,7 +458,7 @@ pub struct MockListener {
 }
 
 #[async_trait::async_trait]
-impl ListenerT for MockListener {
+impl Listener for MockListener {
     type Connection = Connection;
     type Error = std::io::Error;
 
@@ -374,224 +508,30 @@ impl UnreliableConnectionMaker {
 
     pub async fn run(&mut self, connections_end_after: Option<usize>) {
         loop {
-            info!(target: "validator-network", "UnreliableConnectionMaker: waiting for new request...");
+            info!(
+                target: LOG_TARGET,
+                "UnreliableConnectionMaker: waiting for new request..."
+            );
             let (dialer_address, listener_address, c) =
                 self.dialers.next().await.expect("should receive");
-            info!(target: "validator-network", "UnreliableConnectionMaker: received request");
+            info!(
+                target: LOG_TARGET,
+                "UnreliableConnectionMaker: received request"
+            );
             let (dialer_stream, listener_stream) = Connection::new(
                 4096,
                 connections_end_after,
                 dialer_address,
                 listener_address,
             );
-            info!(target: "validator-network", "UnreliableConnectionMaker: sending stream");
+            info!(
+                target: LOG_TARGET,
+                "UnreliableConnectionMaker: sending stream"
+            );
             c.send(dialer_stream).expect("should send");
             self.listeners[listener_address as usize]
                 .unbounded_send(listener_stream)
                 .expect("should send");
         }
     }
-}
-
-#[derive(Clone)]
-struct MockData {
-    data: u32,
-    filler: Vec<u8>,
-    decodes: bool,
-}
-
-impl MockData {
-    fn new(data: u32, filler_size: usize, decodes: bool) -> MockData {
-        MockData {
-            data,
-            filler: vec![0; filler_size],
-            decodes,
-        }
-    }
-}
-
-impl Encode for MockData {
-    fn size_hint(&self) -> usize {
-        self.data.size_hint() + self.filler.size_hint() + self.decodes.size_hint()
-    }
-
-    fn encode_to<T: Output + ?Sized>(&self, dest: &mut T) {
-        // currently this is exactly the default behaviour, but we still
-        // need it here to make sure that decode works in the future
-        self.data.encode_to(dest);
-        self.filler.encode_to(dest);
-        self.decodes.encode_to(dest);
-    }
-}
-
-impl Decode for MockData {
-    fn decode<I: codec::Input>(value: &mut I) -> Result<Self, codec::Error> {
-        let data = u32::decode(value)?;
-        let filler = Vec::<u8>::decode(value)?;
-        let decodes = bool::decode(value)?;
-        if !decodes {
-            return Err("Simulated decode failure.".into());
-        }
-        Ok(Self {
-            data,
-            filler,
-            decodes,
-        })
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn spawn_peer(
-    secret_key: MockSecretKey,
-    addr: Addresses,
-    n_msg: usize,
-    large_message_interval: Option<usize>,
-    corrupted_message_interval: Option<usize>,
-    dialer: MockDialer,
-    listener: MockListener,
-    report: mpsc::UnboundedSender<(MockPublicKey, usize)>,
-    spawn_handle: SpawnTaskHandle,
-) {
-    let our_id = secret_key.public_key();
-    let (service, mut interface) = Service::new(dialer, listener, secret_key, spawn_handle);
-    // run the service
-    tokio::spawn(async {
-        let (_exit, rx) = oneshot::channel();
-        service.run(rx).await;
-    });
-    // start connecting with the peers
-    let mut peer_ids = Vec::with_capacity(addr.len());
-    for (id, addrs) in addr.into_iter() {
-        interface.add_connection(id.clone(), addrs);
-        peer_ids.push(id);
-    }
-    // peer main loop
-    // we send random messages to random peers
-    // a message is a number in range 0..n_msg
-    // we also keep a list of messages received at least once
-    // on receiving a message we report the total number of distinct messages received so far
-    // the goal is to receive every message at least once
-    tokio::spawn(async move {
-        let mut received: HashSet<usize> = HashSet::with_capacity(n_msg);
-        let mut send_ticker = tokio::time::interval(Duration::from_millis(5));
-        let mut counter: usize = 0;
-        loop {
-            tokio::select! {
-                _ = send_ticker.tick() => {
-                    counter += 1;
-                    // generate random message
-                    let filler_size = match large_message_interval {
-                        Some(lmi) if counter % lmi == 0 => TWICE_MAX_DATA_SIZE,
-                        _ => 0,
-                    };
-                    let decodes = match corrupted_message_interval {
-                        Some(cmi) if counter % cmi == 0 => false,
-                        _ => true,
-                    };
-                    let data: MockData = MockData::new(thread_rng().gen_range(0..n_msg) as u32, filler_size, decodes);
-                    // choose a peer
-                    let peer: MockPublicKey = peer_ids[thread_rng().gen_range(0..peer_ids.len())].clone();
-                    // send
-                    interface.send(data, peer);
-                },
-                data = interface.next() => {
-                    // receive the message
-                    let data: MockData = data.expect("next should not be closed");
-                    // mark the message as received, we do not care about sender's identity
-                    received.insert(data.data as usize);
-                    // report the number of received messages
-                    report.unbounded_send((our_id.clone(), received.len())).expect("should send");
-                },
-            };
-        }
-    });
-}
-
-/// Takes O(n log n) rounds to finish, where n = n_peers * n_msg.
-pub async fn scenario(
-    n_peers: usize,
-    n_msg: usize,
-    broken_connection_interval: Option<usize>,
-    large_message_interval: Option<usize>,
-    corrupted_message_interval: Option<usize>,
-    status_report_interval: Duration,
-) {
-    // create spawn_handle, we need to keep the task_manager
-    let task_manager =
-        TaskManager::new(Handle::current(), None).expect("should create TaskManager");
-    let spawn_handle = task_manager.spawn_handle();
-    // create peer identities
-    info!(target: "validator-network", "generating keys...");
-    let keys = random_keys(n_peers);
-    info!(target: "validator-network", "done");
-    // prepare and run the manager
-    let (mut connection_manager, mut callers, addr) =
-        UnreliableConnectionMaker::new(keys.keys().cloned().collect());
-    tokio::spawn(async move {
-        connection_manager.run(broken_connection_interval).await;
-    });
-    // channel for receiving status updates from spawned peers
-    let (tx_report, mut rx_report) = mpsc::unbounded::<(MockPublicKey, usize)>();
-    let mut reports: BTreeMap<MockPublicKey, usize> =
-        keys.keys().cloned().map(|id| (id, 0)).collect();
-    // spawn peers
-    for (id, secret_key) in keys.into_iter() {
-        let mut addr = addr.clone();
-        // do not connect with itself
-        addr.remove(&secret_key.public_key());
-        let (dialer, listener) = callers.remove(&id).expect("should contain all ids");
-        spawn_peer(
-            secret_key,
-            addr,
-            n_msg,
-            large_message_interval,
-            corrupted_message_interval,
-            dialer,
-            listener,
-            tx_report.clone(),
-            spawn_handle.clone(),
-        );
-    }
-    let mut status_ticker = interval(status_report_interval);
-    loop {
-        tokio::select! {
-            maybe_report = rx_report.next() => match maybe_report {
-                Some((peer_id, peer_n_msg)) => {
-                    reports.insert(peer_id, peer_n_msg);
-                    if reports.values().all(|&x| x == n_msg) {
-                        info!(target: "validator-network", "Peers received {:?} messages out of {}, finishing.", reports.values(), n_msg);
-                        return;
-                    }
-                },
-                None => panic!("should receive"),
-            },
-            _ = status_ticker.tick() => {
-                info!(target: "validator-network", "Peers received {:?} messages out of {}.", reports.values(), n_msg);
-            }
-        };
-    }
-}
-
-/// Takes O(n log n) rounds to finish, where n = n_peers * n_msg.
-pub async fn scenario_with_timeout(
-    n_peers: usize,
-    n_msg: usize,
-    broken_connection_interval: Option<usize>,
-    large_message_interval: Option<usize>,
-    corrupted_message_interval: Option<usize>,
-    status_report_interval: Duration,
-    scenario_timeout: Duration,
-) -> Result<(), Elapsed> {
-    timeout(
-        scenario_timeout,
-        scenario(
-            n_peers,
-            n_msg,
-            broken_connection_interval,
-            large_message_interval,
-            corrupted_message_interval,
-            status_report_interval,
-        ),
-    )
-    .await
 }
