+++ finality-aleph/src/network/clique/protocols/v0/mod.rs
@@ -2,13 +2,13 @@ use futures::{channel::mpsc, StreamExt};
 use log::{debug, info, trace};
 use tokio::io::{AsyncRead, AsyncWrite};
 
-use crate::validator_network::{
+use crate::network::clique::{
     io::{receive_data, send_data},
     protocols::{
         handshake::{v0_handshake_incoming, v0_handshake_outgoing},
         ConnectionType, ProtocolError, ResultForService,
     },
-    Data, PublicKey, SecretKey, Splittable,
+    Data, PublicKey, SecretKey, Splittable, LOG_TARGET,
 };
 
 mod heartbeat;
@@ -38,9 +38,12 @@ pub async fn outgoing<SK: SecretKey, D: Data, S: Splittable>(
     public_key: SK::PublicKey,
     result_for_parent: mpsc::UnboundedSender<ResultForService<SK::PublicKey, D>>,
 ) -> Result<(), ProtocolError<SK::PublicKey>> {
-    trace!(target: "validator-network", "Extending hand to {}.", public_key);
+    trace!(target: LOG_TARGET, "Extending hand to {}.", public_key);
     let (sender, receiver) = v0_handshake_outgoing(stream, secret_key, public_key.clone()).await?;
-    info!(target: "validator-network", "Outgoing handshake with {} finished successfully.", public_key);
+    info!(
+        target: LOG_TARGET,
+        "Outgoing handshake with {} finished successfully.", public_key
+    );
     let (data_for_network, data_from_user) = mpsc::unbounded();
     result_for_parent
         .unbounded_send((
@@ -53,7 +56,10 @@ pub async fn outgoing<SK: SecretKey, D: Data, S: Splittable>(
     let sending = sending(sender, data_from_user);
     let heartbeat = heartbeat_receiver(receiver);
 
-    debug!(target: "validator-network", "Starting worker for sending to {}.", public_key);
+    debug!(
+        target: LOG_TARGET,
+        "Starting worker for sending to {}.", public_key
+    );
     loop {
         tokio::select! {
             _ = heartbeat => return Err(ProtocolError::CardiacArrest),
@@ -85,9 +91,12 @@ pub async fn incoming<SK: SecretKey, D: Data, S: Splittable>(
     result_for_parent: mpsc::UnboundedSender<ResultForService<SK::PublicKey, D>>,
     data_for_user: mpsc::UnboundedSender<D>,
 ) -> Result<(), ProtocolError<SK::PublicKey>> {
-    trace!(target: "validator-network", "Waiting for extended hand...");
+    trace!(target: LOG_TARGET, "Waiting for extended hand...");
     let (sender, receiver, public_key) = v0_handshake_incoming(stream, secret_key).await?;
-    info!(target: "validator-network", "Incoming handshake with {} finished successfully.", public_key);
+    info!(
+        target: LOG_TARGET,
+        "Incoming handshake with {} finished successfully.", public_key
+    );
 
     let (tx_exit, mut exit) = mpsc::unbounded();
     result_for_parent
@@ -101,7 +110,10 @@ pub async fn incoming<SK: SecretKey, D: Data, S: Splittable>(
     let receiving = receiving(receiver, data_for_user);
     let heartbeat = heartbeat_sender(sender);
 
-    debug!(target: "validator-network", "Starting worker for receiving from {}.", public_key);
+    debug!(
+        target: LOG_TARGET,
+        "Starting worker for receiving from {}.", public_key
+    );
     loop {
         tokio::select! {
             _ = heartbeat => return Err(ProtocolError::CardiacArrest),
@@ -119,7 +131,7 @@ mod tests {
     };
 
     use super::{incoming, outgoing, ProtocolError};
-    use crate::validator_network::{
+    use crate::network::clique::{
         mock::{key, MockPublicKey, MockSecretKey, MockSplittable},
         protocols::{ConnectionType, ResultForService},
         Data,
