+++ finality-aleph/src/testing/mod.rs
@@ -1,6 +1,6 @@
 pub mod client_chain_builder;
+mod clique_network;
 mod data_store;
 mod justification;
 pub mod mocks;
 mod network;
-mod validator_network;
