+++ scripts/run_nodes.sh
@@ -101,7 +101,7 @@ run_node() {
     --validator-port ${validator_port} \
     -laleph-party=debug \
     -laleph-network=debug \
-    -lvalidator-network=debug \
+    -lclique-network=debug \
     -laleph-finality=debug \
     -laleph-justification=debug \
     -laleph-data-store=debug \
